Remove gRPC create operations
